requirements.txt
git+https://github.com/SENERGY-Platform/client-connector-lib.git@master
python-dotenv==1.0.0